make build shm deletion safer out of paranoia

+++ custom-recipes/build.sh
@@ -3,7 +3,6 @@
 set -euo pipefail
 shopt -s inherit_errexit
 
-
 if [ $# -eq 0 ]; then
   p="./recipes"
 elif [ $# -eq 1 ]; then
@@ -14,26 +13,32 @@ else
 fi
 
 cleanup_tmp_shm() {
-    if [ -d "$tmp_shm" ]; then
-        echo "Cleaning up temporary directory: $tmp_shm"
-        rm -rf "$tmp_shm"
-    fi
+  if [ -n "${tmp_shm:-}" ] && [ -d "$tmp_shm" ]; then
+    echo "Cleaning up temporary directory: $tmp_shm"
+    rm -rf "$tmp_shm"
+  fi
 }
 
 trap cleanup_tmp_shm EXIT
 
+tmp_shm=""
 tmp_shm=$(mktemp -d -p /dev/shm boa-XXXX)
+
+if [ -z "$tmp_shm" ]; then
+  echo "Failed to create temporary directory."
+  exit 1
+fi
 
 conda index ./conda-bld
 
 rm -f ./conda-bld/*/$p-*.tar.bz2
 
 boa build \
-  --numpy 1.24 \
-  --python 3.9 \
-  --croot $tmp_shm \
-  --output-folder ./conda-bld \
-  --target-platform linux-64 \
+  --python=3.9 \
+  --numpy=1.24 \
+  --target-platform=linux-64 \
+  --croot=$tmp_shm \
+  --output-folder=./conda-bld \
   $p
 
 conda index ./conda-bld
